Update purge.md
Data purging is not only about deleting data, but removing the data. The data can be removed to another repository or deleted permanently. It is important to mention that the strategy presented on this article is only about deleting, but due the relevancy of the data, they may move to another permanent location than deleting.